Properly merge always_run when updating jobs
The always_run field in configs/ would get copied to new jobs, but
would be ignored when updating existing jobs.

In theory, since you can specify always_run in configs/ now, we ought
to just always have that value override the existing value in jobs/.
But most existing `always_run: false` jobs don't currently have
`always_run: false` in their config, so that would require a lot of
config fixing first. So we merge them by saying that if _either_ the
config or the job has `always_run: false`, then we make the job have
`always_run: false`. (At this point, we can't distinguish the case
where the config explicitly had `always_run: true` from the case where
that was just filled in as a default.)